restructure project
This commit includes the following quality-of-life improvements:
- switch from makefile to cmake
- move all source and header files to a newly created src/ directory
- add the build directory to .gitignore
- add the .vscode directory to .gitignore

Building the project with cmake can be done in the following way:
    mkdir build && cd build
    cmake ..
    make -j4
Here the first two steps, are only necessary once, subsequents build
are just done through make, like you are already used to.

This commit is heavily oponionated. If you do not like CMake or the new
project structure, feel free to reject.